Rename hooks/ to composables/
