+++ src/components/dialog/content/setting/ServerConfigPanel.vue
@@ -74,8 +74,8 @@ import { watch } from 'vue'
 import { useI18n } from 'vue-i18n'
 
 import FormItem from '@/components/common/FormItem.vue'
+import { useCopyToClipboard } from '@/composables/clipboardHooks'
 import type { ServerConfig } from '@/constants/serverConfig'
-import { useCopyToClipboard } from '@/hooks/clipboardHooks'
 import { useServerConfigStore } from '@/stores/serverConfigStore'
 import { useSettingStore } from '@/stores/settingStore'
 import type { FormItem as FormItemType } from '@/types/settingTypes'
