+++ src/components/graph/GraphCanvas.vue
@@ -41,11 +41,11 @@ import TitleEditor from '@/components/graph/TitleEditor.vue'
 import NodeSearchboxPopover from '@/components/searchbox/NodeSearchBoxPopover.vue'
 import SideToolbar from '@/components/sidebar/SideToolbar.vue'
 import SecondRowWorkflowTabs from '@/components/topbar/SecondRowWorkflowTabs.vue'
+import { useCanvasDrop } from '@/composables/canvasDropHooks'
+import { useGlobalLitegraph } from '@/composables/litegraphHooks'
+import { useContextMenuTranslation } from '@/composables/translationHooks'
+import { useWorkflowPersistence } from '@/composables/workflowPersistenceHooks'
 import { CORE_SETTINGS } from '@/constants/coreSettings'
-import { useCanvasDrop } from '@/hooks/canvasDropHooks'
-import { useGlobalLitegraph } from '@/hooks/litegraphHooks'
-import { useContextMenuTranslation } from '@/hooks/translationHooks'
-import { useWorkflowPersistence } from '@/hooks/workflowPersistenceHooks'
 import { i18n } from '@/i18n'
 import { api } from '@/scripts/api'
 import { app as comfyApp } from '@/scripts/app'
