+++ src/composables/bottomPanelTabs/useTerminal.ts
