+++ src/composables/bottomPanelTabs/useTerminalBuffer.ts
