+++ src/composables/bottomPanelTabs/useTerminalTabs.ts
