+++ src/composables/errorHooks.ts
