+++ src/stores/commandStore.ts
@@ -1,7 +1,7 @@
 import { defineStore } from 'pinia'
 import { computed, ref } from 'vue'
 
-import { useErrorHandling } from '@/hooks/errorHooks'
+import { useErrorHandling } from '@/composables/errorHooks'
 import type { ComfyExtension } from '@/types/comfy'
 
 import { type KeybindingImpl, useKeybindingStore } from './keybindingStore'
