+++ src/stores/workspace/sidebarTabStore.ts
@@ -1,10 +1,10 @@
 import { defineStore } from 'pinia'
 import { computed, ref } from 'vue'
 
-import { useModelLibrarySidebarTab } from '@/hooks/sidebarTabs/modelLibrarySidebarTab'
-import { useNodeLibrarySidebarTab } from '@/hooks/sidebarTabs/nodeLibrarySidebarTab'
-import { useQueueSidebarTab } from '@/hooks/sidebarTabs/queueSidebarTab'
-import { useWorkflowsSidebarTab } from '@/hooks/sidebarTabs/workflowsSidebarTab'
+import { useModelLibrarySidebarTab } from '@/composables/sidebarTabs/modelLibrarySidebarTab'
+import { useNodeLibrarySidebarTab } from '@/composables/sidebarTabs/nodeLibrarySidebarTab'
+import { useQueueSidebarTab } from '@/composables/sidebarTabs/queueSidebarTab'
+import { useWorkflowsSidebarTab } from '@/composables/sidebarTabs/workflowsSidebarTab'
 import { useCommandStore } from '@/stores/commandStore'
 import { SidebarTabExtension } from '@/types/extensionTypes'
