+++ src/views/GraphView.vue
@@ -22,9 +22,9 @@ import UnloadWindowConfirmDialog from '@/components/dialog/UnloadWindowConfirmDi
 import GraphCanvas from '@/components/graph/GraphCanvas.vue'
 import GlobalToast from '@/components/toast/GlobalToast.vue'
 import TopMenubar from '@/components/topbar/TopMenubar.vue'
+import { useCoreCommands } from '@/composables/coreCommandHooks'
+import { useErrorHandling } from '@/composables/errorHooks'
 import { SERVER_CONFIG_ITEMS } from '@/constants/serverConfig'
-import { useCoreCommands } from '@/hooks/coreCommandHooks'
-import { useErrorHandling } from '@/hooks/errorHooks'
 import { i18n } from '@/i18n'
 import { api } from '@/scripts/api'
 import { app } from '@/scripts/app'
