+++ src/views/MaintenanceView.vue
@@ -104,8 +104,8 @@ import BaseTerminal from '@/components/bottomPanel/tabs/terminal/BaseTerminal.vu
 import RefreshButton from '@/components/common/RefreshButton.vue'
 import StatusTag from '@/components/maintenance/StatusTag.vue'
 import TaskListPanel from '@/components/maintenance/TaskListPanel.vue'
-import type { useTerminal } from '@/hooks/bottomPanelTabs/useTerminal'
-import { useTerminalBuffer } from '@/hooks/bottomPanelTabs/useTerminalBuffer'
+import type { useTerminal } from '@/composables/bottomPanelTabs/useTerminal'
+import { useTerminalBuffer } from '@/composables/bottomPanelTabs/useTerminalBuffer'
 import { t } from '@/i18n'
 import { useMaintenanceTaskStore } from '@/stores/maintenanceTaskStore'
 import { MaintenanceFilter } from '@/types/desktop/maintenanceTypes'
